Remove option -r and add more test.
The reason I remove -r is I want the user be fully aware of what is embeded.
The issue with -r is it might embed file silently, and also, when the user will
be able to define a key for each file, how the user will define keys for files not explicitly added?